requirements.txt
PyPDF2
flask
openai
langchain
langchain_community
python-dotenv
